refactor(#5015): extract runtime mismatch safety path
